+++ experimental/packages/otlp-grpc-exporter-base/test/grpc-exporter-transport.test.ts
@@ -136,9 +136,30 @@ describe('GrpcExporterTransport', function () {
     });
   });
   describe('shutdown', function () {
+    let shutdownHandle: () => void | undefined;
+    const serverTestContext: ServerTestContext = {
+      requests: [],
+      serverResponseProvider: () => {
+        return { error: null, buffer: Buffer.from([]) };
+      },
+    };
+
+    beforeEach(async function () {
+      shutdownHandle = await startServer(serverTestContext);
+    });
+
     afterEach(function () {
+      shutdownHandle();
+
+      // clear context
+      serverTestContext.requests = [];
+      serverTestContext.serverResponseProvider = () => {
+        return { error: null, buffer: Buffer.from([]) };
+      };
+
       sinon.restore();
     });
+
     it('before send() does not error', function () {
       const transport = new GrpcExporterTransport(simpleClientConfig);
       transport.shutdown();
@@ -146,27 +167,18 @@ describe('GrpcExporterTransport', function () {
       // no assertions, just checking that it does not throw any errors.
     });
 
-    it('calls client shutdown if client is defined', function () {
-      // arrange
-      const transport = new GrpcExporterTransport({
-        metadata: createEmptyMetadata,
-        timeoutMillis: 100,
-        grpcPath: 'path',
-        grpcName: 'name',
-        credentials: createInsecureCredentials,
-        compression: 'gzip',
-        address: 'localhost:1234',
-      });
-      const shutdownStub = sinon.stub();
-      transport['_client'] = {
-        shutdown: shutdownStub,
-      };
+    it('calls _client.close() if client is defined', async function () {
+      const transport = new GrpcExporterTransport(simpleClientConfig);
+      // send something so that client is defined
+      await transport.send(Buffer.from([1, 2, 3]));
+      assert.ok(transport['_client'], '_client is not defined after send()');
+      const closeSpy = sinon.spy(transport['_client'], 'close');
 
       // act
       transport.shutdown();
 
       // assert
-      sinon.assert.calledOnce(shutdownStub);
+      sinon.assert.calledOnce(closeSpy);
     });
   });
   describe('send', function () {