+++ experimental/CHANGELOG.md
@@ -19,6 +19,8 @@ All notable changes to experimental packages in this project will be documented 
 
 ### :bug: (Bug Fix)
 
+* fix(otlp-grpc-exporter-base): avoid TypeError on exporter shutdown [#4612](https://github.com/open-telemetry/opentelemetry-js/pull/4612)
+
 ### :books: (Refine Doc)
 
 ### :house: (Internal)